Delete old code (#441)
This PR removes the old code after the refactor. There are some todo's in the tests that will be covered by a separate PR
